i66: Prevent multiple renders at the screen level
