+++ src/library/components/index.js
@@ -1,10 +1,12 @@
+export { Menu, MenuDivider, MenuItem } from './Menu';
+export { default as ActivityOverlay } from './ActivityOverlay';
 export { default as FAB } from './FAB';
 export { default as HomeBar } from './HomeBar';
 export { default as IconButton } from './IconButton';
-export { Menu, MenuDivider, MenuItem } from './Menu';
+export { default as Modal } from './Modal';
 export { default as NavBar } from './NavBar';
+export { default as OnboardingButtons } from './OnboardingButtons';
 export { default as Scrim } from './Scrim';
 export { default as TwitterAuth } from './TwitterAuth';
 export { default as WorldMap } from './WorldMap';
-export { default as ActivityOverlay } from './ActivityOverlay';
